requirements.txt
requests
pywebview
pywin32
keyboard
ntplib